Move lib directory to package name
