+++ dakara_player_vlc/daemon.py
