+++ dakara_player_vlc/dakara_manager.py
