+++ dakara_player_vlc/font_loader.py
