+++ dakara_player_vlc/text_generator.py
